+++ tests/cpp/core/CommandLineOptionsTests.cpp
@@ -60,6 +60,7 @@ BOOST_AUTO_TEST_CASE(testCommandLineDefaults)
     BOOST_CHECK_EQUAL(options.sessionsDir, "");
     BOOST_CHECK_EQUAL(options.webservicePort, 0);
     BOOST_CHECK_EQUAL(options.demoServiceUrl, "");
+    BOOST_CHECK_EQUAL(options.targetFPS, 0);
     BOOST_CHECK_EQUAL(options.showPowerButton, false);
     BOOST_CHECK_EQUAL(options.saveDir, "");
 
@@ -76,6 +77,7 @@ void setOptionParameters(CommandLineOptions& options)
     options.sessionsDir = "/var/sessions";
     options.webservicePort = 8888;
     options.demoServiceUrl = "http://demoservice.bbp.epfl.ch";
+    options.targetFPS = 100;
     options.showPowerButton = true;
     options.saveDir = "/tmp/whiteboards";
 }
@@ -91,6 +93,7 @@ void checkOptionParameters(const CommandLineOptions& options)
     BOOST_CHECK_EQUAL(options.sessionsDir, "/var/sessions");
     BOOST_CHECK_EQUAL(options.webservicePort, 8888);
     BOOST_CHECK_EQUAL(options.demoServiceUrl, "http://demoservice.bbp.epfl.ch");
+    BOOST_CHECK_EQUAL(options.targetFPS, 100);
     BOOST_CHECK_EQUAL(options.showPowerButton, true);
     BOOST_CHECK_EQUAL(options.saveDir, "/tmp/whiteboards");
 
@@ -102,6 +105,7 @@ void checkOptionParameters(const CommandLineOptions& options)
                       "--contentsDir /var/contents "
                       "--sessionsDir /var/sessions "
                       "--webservicePort 8888 "
+                      "--targetFPS 100 "
                       "--demoServiceUrl http://demoservice.bbp.epfl.ch "
                       "--showPowerButton "
                       "--saveDir /tmp/whiteboards");