+++ tide/master/rest/FileBrowser.cpp
@@ -1,5 +1,5 @@
 /*********************************************************************/
-/* Copyright (c) 2017, EPFL/Blue Brain Project                       */
+/* Copyright (c) 2017-2019, EPFL/Blue Brain Project                  */
 /*                     Pawel Podhajski <pawel.podhajski@epfl.ch>     */
 /*                     Raphael Dumusc <raphael.dumusc@epfl.ch>       */
 /* All rights reserved.                                              */
@@ -40,9 +40,13 @@
 
 #include "FileBrowser.h"
 
+#include "scene/ContentFactory.h"
 #include "json/json.h"
 
+#include <QDateTime>
 #include <QDir>
+#include <QDirIterator>
+#include <QJsonArray>
 #include <QUrl>
 
 namespace
@@ -59,11 +63,18 @@ QJsonArray _toJsonArray(const QFileInfoList& list)
         array.append(_toJsonObject(entry));
     return array;
 }
+
+bool compareByModificationdDate(const QFileInfo& d1, const QFileInfo& d2)
+{
+    return d1.lastModified() > d2.lastModified();
 }
+
+} // namespace
 
 FileBrowser::FileBrowser(const QString& baseDir, const QStringList& filters)
     : _baseDir{baseDir}
     , _filters{filters}
+    , _extensions{ContentFactory::getSupportedExtensions()}
 {
 }
 
@@ -86,6 +97,82 @@ std::future<rockets::http::Response> FileBrowser::list(
 
     const auto body = json::dump(_toJsonArray(_contents(fullpath)));
     return make_ready_response(Code::OK, body, "application/json");
+}
+
+std::future<rockets::http::Response> FileBrowser::find(
+    const rockets::http::Request& request)
+{
+    using namespace rockets::http;
+    auto path = QString::fromStdString(request.path);
+    QUrl url;
+    url.setPath(path, QUrl::StrictMode);
+    path = url.path();
+    auto queryParam = request.query;
+    const QString fullpath = _baseDir + "/" + path;
+    const QDir absolutePath(fullpath);
+
+    if (!absolutePath.canonicalPath().startsWith(_baseDir))
+        return make_ready_response(Code::BAD_REQUEST);
+
+    if (!absolutePath.exists())
+        return make_ready_response(Code::NO_CONTENT);
+
+    if (queryParam.find("file") == queryParam.end())
+    {
+        return make_ready_response(Code::BAD_REQUEST);
+    }
+
+    auto fileName = QString::fromStdString(queryParam.at("file"));
+    if (fileName.length() < 3)
+        return make_ready_response(Code::BAD_REQUEST);
+
+    const QString fileNameRegex = "*" + fileName + "*";
+
+    auto isSupported = [extensions = _extensions](const QFileInfo& file) {
+        if (file.isDir())
+            return true;
+        for (auto extension : extensions)
+        {
+            if (file.fileName().endsWith(extension))
+                return true;
+        }
+        return false;
+    };
+
+    auto future =
+        std::async(std::launch::async, [isSupported, baseDir = QDir(_baseDir),
+                                        fullpath, fileNameRegex]() {
+            QDirIterator it(fullpath, QStringList() << fileNameRegex,
+                            QDir::Files, QDirIterator::Subdirectories);
+
+            QFileInfoList fileInfoList;
+            while (it.hasNext())
+            {
+                QFileInfo file(it.next());
+
+                if (!isSupported(file))
+                    continue;
+                fileInfoList << file;
+            }
+            std::sort(fileInfoList.begin(), fileInfoList.end(),
+                      compareByModificationdDate);
+
+            QJsonArray list;
+            for (const auto& file : fileInfoList)
+            {
+                QJsonObject obj;
+                obj.insert("name", file.fileName());
+                obj.insert("path",
+                           baseDir.relativeFilePath(file.absoluteFilePath()));
+                obj.insert("size", file.size());
+                obj.insert("isDir", file.isDir());
+                obj.insert("lastModified", file.lastModified().toString());
+                list.append(obj);
+            }
+            const auto body = json::dump(list);
+            return Response(Code::OK, body, "application/json");
+        });
+    return future;
 }
 
 QFileInfoList FileBrowser::_contents(const QDir& directory) const